requirements.txt
psycopg2-binary
python-dotenv
pgvector
databricks-sdk
requests
python-docx
